+++ docs/app/layout.tsx
@@ -1,4 +1,5 @@
 import "@/app/global.css";
+
 import { baseOptions } from "@/app/layout.config";
 import { source } from "@/lib/source";
 import { DocsLayout } from "fumadocs-ui/layouts/docs";