+++ docs/mdx-components.tsx
@@ -1,10 +1,9 @@
-import { ComponentPreview } from "@/components/ComponentPreview";
-import { Sandbox } from "@/components/Sandbox";
 import * as TabsComponents from "fumadocs-ui/components/tabs";
 import * as TypeTableComponents from "fumadocs-ui/components/type-table";
 import defaultMdxComponents from "fumadocs-ui/mdx";
 import * as icons from "lucide-react";
 import type { MDXComponents } from "mdx/types";
+import { LiveComponent } from "./components/LiveComponent";
 
 // use this function to get MDX components, you will need it for rendering MDX
 export function getMDXComponents(components?: MDXComponents): MDXComponents {
@@ -13,8 +12,7 @@ export function getMDXComponents(components?: MDXComponents): MDXComponents {
     ...defaultMdxComponents,
     ...TabsComponents,
     ...TypeTableComponents,
+    LiveComponent,
     ...components,
-    ComponentPreview,
-    Sandbox,
   };
 }